Adds support for list of layer refereneces file to spfs EnvSpecs
There are now 2 kinds of .spfs.yaml file that can be given to spfs run
as environemnt spec items. This adds parsing and enums to support
both, and moves LiveLayers to their own source code file.
