+++ crates/spfs/src/runtime/mod.rs
@@ -4,6 +4,7 @@
 
 //! Handles the setup and initialization of runtime environments
 
+pub mod live_layer;
 #[cfg(unix)]
 pub mod overlayfs;
 #[cfg(unix)]
@@ -16,18 +17,16 @@ mod storage;
 #[cfg(windows)]
 pub mod winfsp;
 
+pub use live_layer::{BindMount, LiveLayer, LiveLayerContents, SpfsFileApiVersion};
 #[cfg(unix)]
 pub use overlayfs::is_removed_entry;
 pub use storage::{
     makedirs_with_perms,
     Author,
-    BindMount,
     Config,
     Data,
     KeyValuePair,
     KeyValuePairBuf,
-    LiveLayer,
-    LiveLayerFile,
     MountBackend,
     OwnedRuntime,
     Runtime,
