+++ crates/spfs/src/runtime/storage.rs
@@ -7,16 +7,13 @@
 use std::borrow::Cow;
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::env::temp_dir;
-use std::fmt::Display;
 use std::fs::OpenOptions;
-use std::io::Read;
 #[cfg(unix)]
 use std::os::unix::fs::{MetadataExt, PermissionsExt};
 #[cfg(windows)]
 use std::os::windows::fs::MetadataExt;
 use std::path::{Path, PathBuf};
 use std::pin::Pin;
-use std::str::FromStr;
 use std::sync::Arc;
 
 use futures::{Stream, StreamExt, TryStreamExt};
@@ -33,6 +30,7 @@ use crate::env::SPFS_DIR_PREFIX;
 use crate::graph::object::Enum;
 use crate::graph::{Annotation, AnnotationValue, KeyAnnotationValuePair};
 use crate::prelude::*;
+use crate::runtime::LiveLayer;
 use crate::storage::fs::DURABLE_EDITS_DIR;
 use crate::storage::RepositoryHandle;
 use crate::{bootstrap, graph, storage, tracking, Error, Result};
@@ -50,10 +48,6 @@ const SPFS_FILESYSTEM_TMPFS_SIZE: &str = "SPFS_FILESYSTEM_TMPFS_SIZE";
 // For durable parameter of create_runtime()
 #[cfg(test)]
 const TRANSIENT: bool = false;
-
-// For the config file that contains the extra bind mounts
-const LIVE_LAYER_FILE_SUFFIX_YAML: &str = ".spfs.yaml";
-const DEFAULT_LIVE_LAYER_FILENAME: &str = "layer.spfs.yaml";
 
 /// Data type for pairs of annotation keys and values
 pub type KeyValuePair<'a> = (&'a str, &'a str);
@@ -114,231 +108,6 @@ pub struct Status {
     /// An empty command signifies that this runtime is being
     /// used to launch an interactive shell environment
     pub command: Vec<String>,
-}
-
-/// Data needed to bind mount a path onto an /spfs backend that uses
-/// overlayfs.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct BindMount {
-    /// Path to the source dir, or file, to bind mount into /spfs at
-    /// the destination.
-    #[serde(alias = "bind")]
-    pub src: PathBuf,
-    /// Where to attach the dir, or file, inside /spfs
-    pub dest: String,
-}
-
-impl BindMount {
-    /// Checks the bind mount is valid for use in /spfs with the given parent directory
-    pub(crate) fn validate(&self, parent: PathBuf) -> Result<()> {
-        if !self.src.starts_with(parent.clone()) {
-            return Err(Error::String(format!(
-                "Bind mount is not valid: {} is not under the live layer's directory: {}",
-                self.src.display(),
-                parent.display()
-            )));
-        }
-
-        if !self.src.exists() {
-            return Err(Error::String(format!(
-                "Bind mount is not valid: {} does not exist",
-                self.src.display()
-            )));
-        }
-
-        Ok(())
-    }
-}
-
-impl Display for BindMount {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}:{}", self.src.display(), self.dest)
-    }
-}
-
-/// The kinds of contents that can be part of a live layer
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum LiveLayerContents {
-    /// A directory or file that will be bind mounted over /spfs
-    BindMount(BindMount),
-}
-
-/// Supported LiveLayer api versions
-#[derive(Debug, Deserialize, Serialize, Copy, Clone, Eq, PartialEq, strum::Display)]
-pub enum LiveLayerApiVersion {
-    #[serde(rename = "v0/layer")]
-    V0Layer,
-}
-
-/// Data needed to add a live layer onto an /spfs overlayfs.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct LiveLayer {
-    /// The api format version of the live layer data
-    pub api: LiveLayerApiVersion,
-    /// The contents that the live layer will put into /spfs
-    pub contents: Vec<LiveLayerContents>,
-}
-
-impl LiveLayer {
-    /// Returns a list of the BindMounts in this LiveLayer
-    pub fn bind_mounts(&self) -> Vec<BindMount> {
-        self.contents
-            .iter()
-            .map(|c| match c {
-                LiveLayerContents::BindMount(bm) => bm.clone(),
-            })
-            .collect::<Vec<_>>()
-    }
-
-    /// Updates the live layer's contents entries using given parent
-    /// directory. This will error if the resulting paths do not exist.
-    ///
-    /// This should be called before validate()
-    fn set_parent(&mut self, parent: PathBuf) -> Result<()> {
-        let mut new_contents = Vec::new();
-
-        for entry in self.contents.iter() {
-            let new_entry = match entry {
-                LiveLayerContents::BindMount(bm) => {
-                    let full_path = match parent.join(bm.src.clone()).canonicalize() {
-                        Ok(abs_path) => abs_path.clone(),
-                        Err(err) => {
-                            return Err(Error::InvalidPath(parent.join(bm.src.clone()), err))
-                        }
-                    };
-
-                    LiveLayerContents::BindMount(BindMount {
-                        src: full_path,
-                        dest: bm.dest.clone(),
-                    })
-                }
-            };
-
-            new_contents.push(new_entry);
-        }
-        self.contents = new_contents;
-
-        Ok(())
-    }
-
-    /// Validates the live layer's contents are under the given parent
-    /// directory and accessible by the current user.
-    ///
-    /// This should be called after set_parent()
-    fn validate(&self, parent: PathBuf) -> Result<()> {
-        for entry in self.contents.iter() {
-            match entry {
-                LiveLayerContents::BindMount(bm) => bm.validate(parent.clone())?,
-            }
-        }
-        Ok(())
-    }
-
-    /// Sets the live layer's parent directory, which updates its
-    /// contents, and then validates its contents.
-    pub fn set_parent_and_validate(&mut self, parent: PathBuf) -> Result<()> {
-        let abs_parent = match parent.canonicalize() {
-            Ok(abs_path) => abs_path.clone(),
-            Err(err) => return Err(Error::InvalidPath(parent.clone(), err)),
-        };
-
-        self.set_parent(parent.clone())?;
-        self.validate(abs_parent)
-    }
-}
-
-impl Display for LiveLayer {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}:{:?}", self.api, self.contents)
-    }
-}
-
-/// A path to a live layer yaml config file, or the path to a
-/// directory that contains a live layer yaml config file named
-/// 'layer.spfs.yaml'.
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub struct LiveLayerFile {
-    filepath: PathBuf,
-}
-
-impl LiveLayerFile {
-    /// Loads the LiveLayer object from this live layer file
-    pub fn load(&self) -> Result<LiveLayer> {
-        tracing::debug!("Opening live layer file: {self}");
-
-        let file = std::fs::File::open(self.filepath.clone()).map_err(|err| {
-            Error::String(format!(
-                "Failed to open file: {} - {err}",
-                self.filepath.display()
-            ))
-        })?;
-
-        let mut data = String::new();
-        std::io::BufReader::new(file)
-            .read_to_string(&mut data)
-            .map_err(|err| {
-                Error::String(format!(
-                    "Failed to read live layer from file {}: {err}",
-                    self.filepath.display()
-                ))
-            })?;
-
-        let mut layer: LiveLayer = serde_yaml::from_str(&data)?;
-
-        let parent = match self.filepath.parent() {
-            Some(p) => p,
-            None => {
-                return Err(Error::String(
-                    "Cannot have a live layer in the top-level root directory".to_string(),
-                ))
-            }
-        };
-        layer.set_parent_and_validate(parent.to_path_buf())?;
-
-        Ok(layer)
-    }
-
-    pub fn parse<S: AsRef<str>>(spec: S) -> Result<Self> {
-        LiveLayerFile::from_str(spec.as_ref())
-    }
-}
-
-impl FromStr for LiveLayerFile {
-    type Err = crate::Error;
-
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        let path = std::path::Path::new(s);
-
-        if path.is_absolute() {
-            let filepath = if path.is_dir() {
-                path.join(DEFAULT_LIVE_LAYER_FILENAME)
-            } else {
-                // Have to use the path as a string because the live
-                // layer filename suffix we are looking for has more
-                // than one '.' in it.
-                let path_string: String = path.display().to_string();
-                if !path_string.ends_with(LIVE_LAYER_FILE_SUFFIX_YAML) {
-                    return Err(Error::String(format!(
-                        "Invalid: {s} does not have a live layer file suffix: {LIVE_LAYER_FILE_SUFFIX_YAML:}"
-        )));
-                }
-                path.to_path_buf()
-            };
-
-            return Ok(Self { filepath });
-        }
-
-        Err(Error::String(format!(
-            "Invalid: {s} is not an absolute path to a live layer"
-        )))
-    }
-}
-
-impl Display for LiveLayerFile {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.filepath.display())
-    }
 }
 
 /// Configuration parameters for the execution of a runtime
