+++ crates/spfs/src/tracking/mod.rs
@@ -15,7 +15,7 @@ mod tag;
 pub use blob_reader::{BlobRead, BlobReadExt};
 pub use diff::{compute_diff, Diff, DiffMode};
 pub use entry::{Entry, EntryKind};
-pub use env::{EnvSpec, EnvSpecItem, ENV_SPEC_EMPTY, ENV_SPEC_SEPARATOR};
+pub use env::{EnvSpec, EnvSpecItem, SpfsFile, ENV_SPEC_EMPTY, ENV_SPEC_SEPARATOR};
 pub use manifest::{
     compute_manifest,
     BlobHasher,
